one big move from all db related functions into db.py
